% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-pkg-events.r
\docType{data}
\name{active}
\alias{active}
\title{Active Settings}
\format{
An object of class \code{environment} of length 4.
}
\usage{
active
}
\description{
Package environment variable to hold
global level settings.
}
\keyword{datasets}
