% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{apply_active_settings}
\alias{apply_active_settings}
\title{Apply Active Logger Settings}
\usage{
apply_active_settings(settings)
}
\arguments{
\item{settings}{defined in the configuration}
}
\description{
Parses and loads the settings specified
in the logger configuration and ensures
they are active in the environment.
}
\seealso{
Other Configuration: 
\code{\link{config_specification}()},
\code{\link{create_log_levels}()},
\code{\link{display_log_levels}()},
\code{\link{ensure_logger}()},
\code{\link{get_active_settings}()},
\code{\link{load_log_layouts}()},
\code{\link{wipe_logger}()}
}
\concept{Configuration}
