% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level.R
\name{as.character.log_level}
\alias{as.character.log_level}
\title{Get Log Level Name}
\usage{
\method{as.character}{log_level}(x, ...)
}
\arguments{
\item{x}{log level}

\item{...}{ignored}
}
\value{
log level name
}
\description{
gets the name of the log level though
casting to a character and forwarding the call
to get_level_name.
}
