% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level.R
\name{as.integer.log_level}
\alias{as.integer.log_level}
\title{Gets the severity of a log level.}
\usage{
\method{as.integer}{log_level}(x, ...)
}
\arguments{
\item{x}{log level}

\item{...}{ignored}
}
\value{
log level
}
\description{
Gets the severity of a log level.
}
