% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{attach_log_levels}
\alias{attach_log_levels}
\title{Load Log Levels}
\usage{
attach_log_levels(levels)
}
\arguments{
\item{levels}{defined in the configuration}
}
\description{
Parses and loads the levels specified in the
logging configuration and registers them with the
dispatcher via the \code{log_levels} active
binding.
}
\seealso{
Other Logging: 
\code{\link{LogDispatch}},
\code{\link{display_log_levels}()},
\code{\link{load_log_layouts}()},
\code{\link{set_log_configuration}()}
}
\concept{Logging}
