% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{class_scope}
\alias{class_scope}
\title{Calling Class Scope}
\usage{
class_scope(cls)
}
\arguments{
\item{cls}{R6 class to export.}
}
\value{
system context for evaluating \code{fmt_metric} objects.
}
\description{
Gets the exposed public field scope of a R6 class. Used for
evaluating cls field execution scopes.
}
\seealso{
Other Context: 
\code{\link{exec_context}()},
\code{\link{format_fn_call}()},
\code{\link{get_call_stack}()},
\code{\link{get_r_version}()},
\code{\link{get_system_info}()},
\code{\link{sys_context}()}
}
\concept{Context}
