% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_specification}
\alias{config_specification}
\title{Config Specification}
\usage{
config_specification()
}
\value{
Nothing.
}
\description{
Loads & attaches a logger with the specified
config.
}
\seealso{
Other Configuration: 
\code{\link{apply_active_settings}()},
\code{\link{create_log_levels}()},
\code{\link{display_log_levels}()},
\code{\link{ensure_logger}()},
\code{\link{get_active_settings}()},
\code{\link{load_log_layouts}()},
\code{\link{wipe_logger}()}
}
\concept{Configuration}
