% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{configure_logger}
\alias{configure_logger}
\title{Configure Logger}
\usage{
configure_logger(settings)
}
\arguments{
\item{settings}{loaded from config}
}
\description{
Parses and loads settings for the log dispatcher.
}
\seealso{
Other Logging: 
\code{\link{LogDispatch}},
\code{\link{attach_log_levels}()},
\code{\link{display_log_levels}()},
\code{\link{get_configurations}()},
\code{\link{init_logger}()},
\code{\link{load_log_layouts}()},
\code{\link{set_log_configuration}()}
}
\concept{Logging}
