% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{display_log_levels}
\alias{display_log_levels}
\title{Display Log Levels}
\usage{
display_log_levels()
}
\value{
Nothing.
}
\description{
A utility function that dynamically displays the configured
log levels (loaded from config), and outputs them in a simple
layout with only the log level and msg formatted in their
crayon styles.
}
\seealso{
Other Logging: 
\code{\link{LogDispatch}},
\code{\link{attach_log_levels}()},
\code{\link{load_log_layouts}()},
\code{\link{set_log_configuration}()}
}
\concept{Logging}
