% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-pkg-events.r
\name{.onAttach}
\alias{.onAttach}
\title{Attach Handler

Package initialization routine.}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{library name}

\item{pkgname}{package name}
}
\description{
Attach Handler

Package initialization routine.
}
