% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-pkg-events.r
\name{.onDetach}
\alias{.onDetach}
\title{Detach Handler

Package initialization routine.}
\usage{
.onDetach(libpath)
}
\arguments{
\item{libpath}{libpath}
}
\description{
Detach Handler

Package initialization routine.
}
