% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-pkg-events.r
\name{.onLoad}
\alias{.onLoad}
\title{Load Handler

Package initialization routine.}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{library name}

\item{pkgname}{package name}
}
\description{
Load Handler

Package initialization routine.
}
