% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{ensure_instance}
\alias{ensure_instance}
\title{Ensure Instance}
\usage{
ensure_instance(variable)
}
\arguments{
\item{variable}{variable name.}
}
\value{
None.
}
\description{
Ensures there is an active dispatcher
attached to the specified environment.
}
\seealso{
Other Configuration: 
\code{\link{config_specification}()},
\code{\link{display_log_levels}()},
\code{\link{load_log_layouts}()},
\code{\link{wipe_logger}()}
}
\concept{Configuration}
