% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{exec_context}
\alias{exec_context}
\title{Execution Context}
\usage{
exec_context(
  keep_args = FALSE,
  max_calls = 5,
  call_subset = c(-1, -1),
  filter_internal = TRUE
)
}
\arguments{
\item{keep_args}{bool to specify keep function all arguments}

\item{max_calls}{maximum number of calls to keep from the stack}

\item{call_subset}{offset index into system calls}

\item{filter_internal}{filter out internal calls?}
}
\value{
system context for evaluating \code{fmt_metric} objects.
}
\description{
Wrapper around Sys.info() and \code{get_r_version} that provides
a consolidated list of variables used for logging contexts.
}
\seealso{
Other Context: 
\code{\link{class_scope}()},
\code{\link{format_fn_call}()},
\code{\link{get_call_stack}()},
\code{\link{get_r_version}()},
\code{\link{get_system_info}()},
\code{\link{sys_context}()}
}
\concept{Context}
