% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{extract_func_name}
\alias{extract_func_name}
\title{Extract Function Name}
\usage{
extract_func_name(func)
}
\arguments{
\item{func}{function name}
}
\value{
function name without arguments
}
\description{
Extracts the name of the function from a deparse call.
}
\seealso{
Other Context: 
\code{\link{class_scope}()},
\code{\link{clean_internal_calls}()},
\code{\link{exec_context}()},
\code{\link{format_fn_call}()},
\code{\link{get_call_stack}()},
\code{\link{get_r_version}()},
\code{\link{get_system_info}()},
\code{\link{is_logger_call}()},
\code{\link{sys_context}()}
}
\concept{Context}
