% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format.fmt_timestamp}
\alias{format.fmt_timestamp}
\title{Gets the format of a format object.}
\usage{
\method{format}{fmt_timestamp}(x, ...)
}
\arguments{
\item{x}{object to extract value from.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
object's value
}
\description{
Gets the format of a format object.
}
