% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{format.log_layout}
\alias{format.log_layout}
\title{Log Layout Format}
\usage{
\method{format}{log_layout}(fmt, ...)
}
\arguments{
\item{fmt}{object to extract value from.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
layout format
}
\description{
Gets the format of a format object.
}
