% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level.R
\name{format.log_level}
\alias{format.log_level}
\title{Log Level Format}
\usage{
\method{format}{log_level}(x, message = character(0), ...)
}
\arguments{
\item{x}{log level}

\item{message}{message to format}

\item{...}{further arguments passed to or from other methods.}
}
\value{
styled level information
}
\description{
formats a message with the style of the log level.
}
\examples{
\dontrun{
level_info(LEVEL)
}
}
