% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{format_fn_call}
\alias{format_fn_call}
\title{Format Function Call}
\usage{
format_fn_call(expr, cutoff = 100L)
}
\arguments{
\item{expr}{function call}

\item{cutoff}{max width cutoff}
}
\value{
string representation of a func call.
}
\description{
Formats a function call into a deparsed string.
}
\seealso{
Other Context: 
\code{\link{class_scope}()},
\code{\link{exec_context}()},
\code{\link{get_call_stack}()},
\code{\link{get_r_version}()},
\code{\link{get_system_info}()},
\code{\link{sys_context}()}
}
\concept{Context}
