% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_active_settings}
\alias{get_active_settings}
\title{Active Logger Settings}
\usage{
get_active_settings()
}
\description{
Gets the active global settings
for the logger.
}
\seealso{
Other Configuration: 
\code{\link{apply_active_settings}()},
\code{\link{config_specification}()},
\code{\link{create_log_levels}()},
\code{\link{display_log_levels}()},
\code{\link{ensure_logger}()},
\code{\link{load_log_layouts}()},
\code{\link{wipe_logger}()}
}
\concept{Configuration}
