% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_configurations}
\alias{get_configurations}
\title{Get Configurations}
\usage{
get_configurations(pkgname = "dyn.log")
}
\arguments{
\item{pkgname}{package name to get configs for.}
}
\description{
Gets all available logging configurations
exposed by the package.
}
\seealso{
Other Logging: 
\code{\link{LogDispatch}},
\code{\link{init_logger}()}
}
\concept{Logging}
