% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{get_format_types}
\alias{get_format_types}
\title{Log Layout Format Types}
\usage{
get_format_types(layout)
}
\arguments{
\item{layout}{object to extract value from.}
}
\value{
layout format
}
\description{
Gets the distinct format types in a log layout instance,
which is useful for determining the appropriate amount
of log context to construct.
}
