% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{get_layout}
\alias{get_layout}
\title{Get Layout}
\usage{
get_layout(association)
}
\arguments{
\item{association}{associated class}
}
\value{
associated \code{log_layout}
}
\description{
gets a log layout associated with a class.
}
