% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level.R
\name{get_minimum_severity}
\alias{get_minimum_severity}
\title{Get Minimum Severity}
\usage{
get_minimum_severity()
}
\value{
styled level information
}
\description{
Gets the log level with the highest threshold
which is used internally to log 'trace/info'
level messages.
}
