% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{get_r_version}
\alias{get_r_version}
\title{R Version}
\usage{
get_r_version()
}
\value{
R environment version is (major).(minor) format
}
\description{
Wrapper around \code{R.Version()} to produce a nicely
formatted string for use use in sys_context.
}
