% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{get_system_info}
\alias{get_system_info}
\title{Get System Info}
\usage{
get_system_info()
}
\value{
\code{Sys.info()} as a named list
}
\description{
Wrapper around Sys.info() that provides the values
in a named list format.
}
