% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{init_logger}
\alias{init_logger}
\title{Init Logger}
\usage{
init_logger(pos = 1)
}
\arguments{
\item{pos}{enviroment level (1=topenv)}
}
\description{
Attaches a reference to the global
logger to the top env if one doesn't
already exist.
}
\seealso{
Other Logging: 
\code{\link{LogDispatch}},
\code{\link{attach_log_levels}()},
\code{\link{configure_logger}()},
\code{\link{display_log_levels}()},
\code{\link{load_log_layouts}()},
\code{\link{set_log_configuration}()}
}
\concept{Logging}
