% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{is_logger_call}
\alias{is_logger_call}
\title{Is Logger Call}
\usage{
is_logger_call(call)
}
\arguments{
\item{call}{function call}
}
\value{
string representation of a func call.
}
\description{
Determines if a call came from the logger, so we
can exclude it from the call stack.
}
\seealso{
Other Internal: 
\code{\link{clean_internal_calls}()},
\code{\link{extract_func_name}()}
}
\concept{Internal}
