% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{length.log_layout}
\alias{length.log_layout}
\title{Log Layout Length}
\usage{
\method{length}{log_layout}(x, ...)
}
\arguments{
\item{x}{log format}

\item{...}{further arguments passed to or from other methods.}
}
\value{
number of formats in the layout.
}
\description{
Generic override for length of a log layout that returns
the number of individual format objects in the layout.
}
