% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level.R
\name{level_description}
\alias{level_description}
\title{Log Level Description}
\usage{
level_description(level)

level_description(level)
}
\arguments{
\item{level}{log level}
}
\value{
level description

level severity
}
\description{
Gets the description of a log level.

Gets the description of a log level.
}
\examples{
\dontrun{
level_description(LEVEL)
}
\dontrun{
level_description(LEVEL)
}
}
