% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level.R
\name{level_format}
\alias{level_format}
\title{Format}
\usage{
level_format(level, msg)
}
\arguments{
\item{level}{log level}

\item{msg}{log msg}
}
\value{
formatted log message
}
\description{
Gets log level message format.
}
