% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level.R
\name{level_info}
\alias{level_info}
\title{Log Level Information}
\usage{
level_info(level)
}
\arguments{
\item{level}{log level}
}
\value{
log level information
}
\description{
Gets log level information.
}
\examples{
\dontrun{
level_info(LEVEL)
}
}
