% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level.R
\name{level_severity}
\alias{level_severity}
\title{get level severity}
\usage{
level_severity(level)

level_severity(level)
}
\arguments{
\item{level}{log level}
}
\value{
level severity

level severity
}
\description{
Gets the severity of a log level.
}
\examples{
\dontrun{
level_severity(LEVEL)
}
\dontrun{
level_severity(LEVEL)
}
}
