% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_log_layouts}
\alias{load_log_layouts}
\title{Load Log Layouts}
\usage{
load_log_layouts(layouts)
}
\arguments{
\item{layouts}{defined in the configuration}
}
\description{
Parses and loads layouts specified in the logging
configuration and registers them with the log
dispatcher via the \code{log_layouts} active
binding.
}
\seealso{
Other Logging: 
\code{\link{LogDispatch}},
\code{\link{attach_log_levels}()},
\code{\link{configure_logger}()},
\code{\link{display_log_levels}()},
\code{\link{init_logger}()},
\code{\link{set_log_configuration}()}
}
\concept{Logging}
