% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{log_layout_detail}
\alias{log_layout_detail}
\title{Log Layout Detail}
\usage{
log_layout_detail(layout)
}
\arguments{
\item{layout}{object to extract layout detail from.}
}
\value{
layout format
}
\description{
Gets the layout formats and the distinct format types in
a log layout instance, which is useful for determining
the appropriate amount of log context to construct.
}
