% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{new_fmt_cls_field}
\alias{new_fmt_cls_field}
\title{Formatted field from the calling class scope.}
\usage{
new_fmt_cls_field(style, field)
}
\arguments{
\item{style}{{crayon::style()}}

\item{field}{field in the object to display}
}
\value{
\code{new_fmt_cls_field}
}
\description{
Placeholder for a container class field
}
\seealso{
Other Log Layout: 
\code{\link{evaluate_layout}()},
\code{\link{new_fmt_exec_scope}()},
\code{\link{new_fmt_layout}()},
\code{\link{new_fmt_line_break}()},
\code{\link{new_fmt_literal}()},
\code{\link{new_fmt_log_msg}()},
\code{\link{new_fmt_metric}()},
\code{\link{new_fmt_timestamp}()},
\code{\link{new_log_layout}()}
}
\concept{Log Layout}
