% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{new_fmt_layout}
\alias{new_fmt_layout}
\title{Format Layout}
\usage{
new_fmt_layout(style)
}
\arguments{
\item{style}{\link{crayon} that the layout will use in log generation.}
}
\value{
new log format
}
\description{
Base type for log format objects.
}
\seealso{
Other Log Layout: 
\code{\link{evaluate_layout}()},
\code{\link{new_fmt_cls_field}()},
\code{\link{new_fmt_exec_scope}()},
\code{\link{new_fmt_line_break}()},
\code{\link{new_fmt_literal}()},
\code{\link{new_fmt_log_msg}()},
\code{\link{new_fmt_metric}()},
\code{\link{new_fmt_timestamp}()},
\code{\link{new_log_layout}()}
}
\concept{Log Layout}
