% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{new_fmt_line_break}
\alias{new_fmt_line_break}
\title{Formatted Line Break}
\usage{
new_fmt_line_break()
}
\value{
log layout newline.
}
\description{
Placeholder for a new line in a log layout.
}
\seealso{
Other Log Layout: 
\code{\link{evaluate_layout}()},
\code{\link{new_fmt_cls_field}()},
\code{\link{new_fmt_exec_scope}()},
\code{\link{new_fmt_layout}()},
\code{\link{new_fmt_literal}()},
\code{\link{new_fmt_log_msg}()},
\code{\link{new_fmt_metric}()},
\code{\link{new_fmt_timestamp}()},
\code{\link{new_log_layout}()}
}
\concept{Log Layout}
