% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{new_fmt_literal}
\alias{new_fmt_literal}
\title{Formatted Literal}
\usage{
new_fmt_literal(style, literal)
}
\arguments{
\item{style}{format style (crayon)}

\item{literal}{log value}
}
\value{
log metric layout.
}
\description{
Placeholder for a formatted literal in a log layout.
}
\examples{
\dontrun{
new_fmt_literal(red $ bold, "literal text")

new_fmt_literal(blue $ italic, "literal text")
}
}
\seealso{
Other Log Layout: 
\code{\link{evaluate_layout}()},
\code{\link{new_fmt_cls_field}()},
\code{\link{new_fmt_exec_scope}()},
\code{\link{new_fmt_layout}()},
\code{\link{new_fmt_line_break}()},
\code{\link{new_fmt_log_msg}()},
\code{\link{new_fmt_metric}()},
\code{\link{new_fmt_timestamp}()},
\code{\link{new_log_layout}()}
}
\concept{Log Layout}
