% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{new_fmt_log_level}
\alias{new_fmt_log_level}
\title{Formatted Log Level}
\usage{
new_fmt_log_level()
}
\value{
a \code{fmt_log_level}.
}
\description{
Placeholder for the formatted log level in a log layout.
}
\concept{Formats}
