% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{new_fmt_metric}
\alias{new_fmt_metric}
\title{Formatted Metric}
\usage{
new_fmt_metric(style, metric)
}
\arguments{
\item{style}{that the layout will use in log generation}

\item{metric}{the metric to log.}
}
\value{
a new formatted metric
}
\description{
Inserts a formatted log metric.
}
\examples{
\dontrun{
new_fmt_metric(bold $ green, "sysname")

new_fmt_metric(bold $ red, "release")
}
}
\seealso{
\link{LogDispatch}

Other Log Layout: 
\code{\link{evaluate_layout}()},
\code{\link{new_fmt_cls_field}()},
\code{\link{new_fmt_exec_scope}()},
\code{\link{new_fmt_layout}()},
\code{\link{new_fmt_line_break}()},
\code{\link{new_fmt_literal}()},
\code{\link{new_fmt_log_msg}()},
\code{\link{new_fmt_timestamp}()},
\code{\link{new_log_layout}()}
}
\concept{Log Layout}
