% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level.R
\name{new_log_level}
\alias{new_log_level}
\title{Log Level}
\usage{
new_log_level(name, description, severity, log_style = NULL, msg_style = NULL)
}
\arguments{
\item{name}{name of the log level is the string representation.}

\item{description}{description of the log level & limited info
on appropriate usage.}

\item{severity}{log severity is used in determining if a
message should get displayed according to the currently set
evaluation threshold.}

\item{log_style}{is a {crayon::style()} that will colorize
the log level.}

\item{msg_style}{is a {crayon::style()} style that will gray scale the
log message, with typically inverted strength, according to the severity.}
}
\value{
\code{log_level}
}
\description{
S3 object to represent a typed & predefined log level.
}
\concept{Log Level}
