% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_log_configuration}
\alias{set_log_configuration}
\title{Load Log Levels}
\usage{
set_log_configuration(file_name, envir = parent.frame())
}
\arguments{
\item{file_name}{loads logging levels from a yml configuration file.}
}
\value{
initialized log levels from the specified config.
}
\description{
Load Log Levels
}
\seealso{
Other Logging: 
\code{\link{LogDispatch}}
}
\concept{Logging}
