% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{style.fmt_layout}
\alias{style.fmt_layout}
\title{Style}
\usage{
\method{style}{fmt_layout}(obj, ...)
}
\arguments{
\item{obj}{object to extract value from.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
object's value
}
\description{
Gets the style of a format object.
}
