% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level.R
\name{style.log_level}
\alias{style.log_level}
\title{Get Level Style}
\usage{
\method{style}{log_level}(obj, ...)
}
\arguments{
\item{obj}{log level}

\item{...}{further arguments passed to or from other methods.}
}
\value{
log level name
}
\description{
gets the style of the log level.
}
