% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{style}
\alias{style}
\title{Style}
\usage{
style(obj)
}
\arguments{
\item{obj}{object to extract value from.}
}
\value{
object's value
}
\description{
Generic style method used for overriding to
get style information from various logging objects.
}
