% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{sys_context}
\alias{sys_context}
\title{System Context}
\usage{
sys_context()
}
\value{
system context for evaluating \code{fmt_metric} objects.
}
\description{
Wrapper around Sys.info() and \code{get_r_version} that provides
a consolidated list of variables used for logging contexts.
}
\seealso{
Other Context: 
\code{\link{class_scope}()},
\code{\link{clean_internal_calls}()},
\code{\link{exec_context}()},
\code{\link{extract_func_name}()},
\code{\link{format_fn_call}()},
\code{\link{get_call_stack}()},
\code{\link{get_r_version}()},
\code{\link{get_system_info}()},
\code{\link{is_logger_call}()}
}
\concept{Context}
