% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{value.fmt_cls_field}
\alias{value.fmt_cls_field}
\title{Value}
\usage{
\method{value}{fmt_cls_field}(obj, cls_context, ...)
}
\arguments{
\item{obj}{object to extract value from.}

\item{cls_context}{class scope to evaluate with.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
object's value
}
\description{
Generic override for getting the value of an
enclosing class variable.
}
