% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{value.fmt_level_info}
\alias{value.fmt_level_info}
\title{Value}
\usage{
\method{value}{fmt_level_info}(fmt, ...)
}
\arguments{
\item{fmt}{object to extract value from.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
object's value
}
\description{
Generic override for getting the value for
log level information.
}
