% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{value.fmt_literal}
\alias{value.fmt_literal}
\title{Value}
\usage{
\method{value}{fmt_literal}(obj, ...)
}
\arguments{
\item{obj}{object to extract value from.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
object's value
}
\description{
Generic override for getting the value of a
literal log message.
}
