% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{value.fmt_log_msg}
\alias{value.fmt_log_msg}
\title{Value}
\usage{
\method{value}{fmt_log_msg}(obj, msg_context, ...)
}
\arguments{
\item{obj}{object to extract value from.}

\item{msg_context}{context to evaluate log message.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
object's value
}
\description{
Generic override for getting the value of an
log format message.
}
