% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{value.fmt_metric}
\alias{value.fmt_metric}
\title{Value}
\usage{
\method{value}{fmt_metric}(obj, sys_context, ...)
}
\arguments{
\item{obj}{object to extract value from.}

\item{sys_context}{context to evaluate the metric.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
object's value
}
\description{
Generic override for getting the value of an
system info variable.
}
