% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{value.log_layout}
\alias{value.log_layout}
\title{A container for a set of formatters that specify the log
entry layout.}
\usage{
\method{value}{log_layout}(layout, ...)
}
\arguments{
\item{layout}{collection of format objects to initialize with.}

\item{...}{further arguments passed to or from other methods.}

\item{association}{class of objects to associate this format with.}
}
\value{
evaluated log layout
}
\description{
A container for a set of formatters that specify the log
entry layout.
}
\seealso{
Other Log Layout: 
\code{\link{get_call_stack}()},
\code{\link{new_fmt_cls_field}()},
\code{\link{new_fmt_layout}()},
\code{\link{new_fmt_line_break}()},
\code{\link{new_fmt_literal}()},
\code{\link{new_fmt_log_level}()},
\code{\link{new_fmt_log_msg}()},
\code{\link{new_fmt_metric}()},
\code{\link{new_fmt_timestamp}()},
\code{\link{new_log_layout}()}
}
\concept{Log Layout}
