% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{value.new_fmt_cls_field}
\alias{value.new_fmt_cls_field}
\title{Gets the value of a log msg object.}
\usage{
\method{value}{new_fmt_cls_field}(fmt, cls_scope, ...)
}
\arguments{
\item{fmt}{object to extract value from.}

\item{cls_scope}{class scope to evalute with.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
object's value
}
\description{
Gets the value of a log msg object.
}
