% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{value.new_fmt_log_msg}
\alias{value.new_fmt_log_msg}
\title{Value}
\usage{
\method{value}{new_fmt_log_msg}(fmt, ...)
}
\arguments{
\item{fmt}{object to extract value from.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
object's value
}
\description{
Generic override for getting the value of an
log format message.
}
