% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{value}
\alias{value}
\title{Value}
\usage{
value(fmt, ...)
}
\arguments{
\item{fmt}{object to extract value from.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
object's value
}
\description{
Base method for getting the value of a
format object.
}
