% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{wipe_logger}
\alias{wipe_logger}
\title{Wipe the Logger Instance}
\usage{
wipe_logger(name, envir = parent.frame())
}
\arguments{
\item{name}{variable name.}

\item{envir}{environment}
}
\value{
None.
}
\description{
Cleans up any dangling global instance
from a previous load.
}
\seealso{
Other Configuration: 
\code{\link{config_specification}()},
\code{\link{display_log_levels}()},
\code{\link{load_log_layouts}()}
}
\concept{Configuration}
