% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{clean_internal_calls}
\alias{clean_internal_calls}
\title{Clean System Calls}
\usage{
clean_internal_calls(call_stack)
}
\arguments{
\item{call_stack}{call stack}
}
\value{
string representation of a func call.
}
\description{
Cleans up any internal system calls from
inside the package from the call stack.
}
\seealso{
Other Internal: 
\code{\link{extract_func_name}()},
\code{\link{is_logger_call}()}
}
\concept{Internal}
