% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{display_log_levels}
\alias{display_log_levels}
\title{Display Log Levels}
\usage{
display_log_levels()
}
\description{
A utility function that dynamically displays the configured
log levels (loaded from config), and outputs them in a simple
layout with only the log level and msg formatted in their
crayon styles.
}
\seealso{
Other Configuration: 
\code{\link{apply_active_settings}()},
\code{\link{config_specification}()},
\code{\link{create_log_levels}()},
\code{\link{ensure_logger}()},
\code{\link{get_active_settings}()},
\code{\link{load_log_layouts}()},
\code{\link{wipe_logger}()}
}
\concept{Configuration}
