% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{evaluate_layout}
\alias{evaluate_layout}
\title{A container for a set of formatters that specify the log
entry layout.}
\usage{
evaluate_layout(formats, types, seperator, context, new_line = "\\n")
}
\arguments{
\item{formats}{the layout formats needed for evaluation.}

\item{types}{the distinct list of format types in the layout.}

\item{seperator}{the layout separator that is inserted between entries.}

\item{context}{a list of contexts needed to evaluate formats in the
the layout.}

\item{new_line}{the layout separator that is inserted between lines.}
}
\value{
evaluated log layout
}
\description{
A container for a set of formatters that specify the log
entry layout.
}
\seealso{
Other Log Layout: 
\code{\link{get_call_stack}()},
\code{\link{new_fmt_cls_field}()},
\code{\link{new_fmt_exec_scope}()},
\code{\link{new_fmt_layout}()},
\code{\link{new_fmt_line_break}()},
\code{\link{new_fmt_literal}()},
\code{\link{new_fmt_log_level}()},
\code{\link{new_fmt_log_msg}()},
\code{\link{new_fmt_metric}()},
\code{\link{new_fmt_timestamp}()},
\code{\link{new_log_layout}()}
}
\concept{Log Layout}
