% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{evaluate_layout}
\alias{evaluate_layout}
\title{A container for a set of formatters that specify the log
entry layout.}
\usage{
evaluate_layout(layout, context)
}
\arguments{
\item{layout}{The layout specified for evaluation.}

\item{context}{a list of contexts needed to evaluate formats in the
the layout.}
}
\value{
evaluated log layout
}
\description{
A container for a set of formatters that specify the log
entry layout.
}
\seealso{
Other Log Layout: 
\code{\link{new_fmt_cls_field}()},
\code{\link{new_fmt_exec_scope}()},
\code{\link{new_fmt_layout}()},
\code{\link{new_fmt_line_break}()},
\code{\link{new_fmt_literal}()},
\code{\link{new_fmt_log_msg}()},
\code{\link{new_fmt_metric}()},
\code{\link{new_fmt_timestamp}()},
\code{\link{new_log_layout}()}
}
\concept{Log Layout}
