% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{get_call_stack}
\alias{get_call_stack}
\title{Formatted Call Stack}
\usage{
get_call_stack(keep_args = F)
}
\arguments{
\item{keep_args}{boolean to indicate if you wante to keep
call arguments or not.}
}
\value{
formatted call stack
}
\description{
Placeholder for the formatted call stack in a log layout.
}
\seealso{
Other Log Layout: 
\code{\link{evaluate_layout}()},
\code{\link{new_fmt_cls_field}()},
\code{\link{new_fmt_exec_scope}()},
\code{\link{new_fmt_layout}()},
\code{\link{new_fmt_line_break}()},
\code{\link{new_fmt_literal}()},
\code{\link{new_fmt_log_level}()},
\code{\link{new_fmt_log_msg}()},
\code{\link{new_fmt_metric}()},
\code{\link{new_fmt_timestamp}()},
\code{\link{new_log_layout}()}
}
\concept{Log Layout}
