% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{get_system_info}
\alias{get_system_info}
\title{Get System Info}
\usage{
get_system_info()
}
\value{
\code{Sys.info()} as a named list
}
\description{
Wrapper around Sys.info() that provides the values
in a named list format.
}
\seealso{
Other Context: 
\code{\link{class_scope}()},
\code{\link{exec_context}()},
\code{\link{format_fn_call}()},
\code{\link{get_call_stack}()},
\code{\link{get_r_version}()},
\code{\link{sys_context}()}
}
\concept{Context}
