% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{init_logger}
\alias{init_logger}
\title{Init Logger}
\usage{
init_logger(file_path = NULL)
}
\arguments{
\item{file_path}{logging configuration to use.}
}
\description{
Loads the configuration passed in,
or uses the default if nothing is specified,
and attaches a reference to the singleton
dispatcher to the global environment.
}
\seealso{
Other Logging: 
\code{\link{LogDispatch}},
\code{\link{get_configurations}()}
}
\concept{Logging}
