% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_log_layouts}
\alias{load_log_layouts}
\title{Load Log Layouts}
\usage{
load_log_layouts(layouts)
}
\arguments{
\item{layouts}{defined in the configuration}
}
\value{
None.
}
\description{
Parses and loads layouts specified in the logging
configuration and registers them with the log
dispatcher via the \code{log_layouts} active
binding.
}
\seealso{
Other Configuration: 
\code{\link{apply_active_settings}()},
\code{\link{config_specification}()},
\code{\link{create_log_levels}()},
\code{\link{display_log_levels}()},
\code{\link{ensure_logger}()},
\code{\link{get_active_settings}()},
\code{\link{wipe_logger}()}
}
\concept{Configuration}
