% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{log_layouts}
\alias{log_layouts}
\title{Log Layouts}
\usage{
log_layouts(association = character(0), layout = NULL)
}
\arguments{
\item{association}{named association to the layout}

\item{layout}{log layout to add if not already existing.}
}
\value{
defined log layouts
}
\description{
an active binding to keep track of log layouts created
with \code{new_log_layout}.
}
