% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{Logger}
\alias{Logger}
\title{Global Logging Instance}
\format{
An object of class \code{LogDispatch} (inherits from \code{R6}) of length 16.
}
\usage{
Logger
}
\description{
Global Logging Instance
}
\keyword{datasets}
