% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{new_log_layout}
\alias{new_log_layout}
\title{Log Layout}
\usage{
new_log_layout(..., sep = " ", association = character())
}
\arguments{
\item{...}{collection of format objects to initialize with.}

\item{sep}{format entry separator, defaults to a single space.}

\item{association}{objects to associate this log format with.}
}
\value{
object's value
}
\description{
a class that stores a collection of log format objects
and understands how to associate a given format to
a class of objects.
}
\seealso{
Other Log Layout: 
\code{\link{evaluate_layout}()},
\code{\link{get_call_stack}()},
\code{\link{new_fmt_cls_field}()},
\code{\link{new_fmt_exec_scope}()},
\code{\link{new_fmt_layout}()},
\code{\link{new_fmt_line_break}()},
\code{\link{new_fmt_literal}()},
\code{\link{new_fmt_log_level}()},
\code{\link{new_fmt_log_msg}()},
\code{\link{new_fmt_metric}()},
\code{\link{new_fmt_timestamp}()}
}
\concept{Log Layout}
