% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_log_configuration}
\alias{set_log_configuration}
\title{Set Logging Configuration}
\usage{
set_log_configuration(file_name, envir = parent.frame())
}
\arguments{
\item{file_name}{loads logging levels from a yml configuration file.}

\item{envir}{environment to load logger into.}
}
\value{
initialized log levels from the specified config.
}
\description{
loads the logging configuration from the
yaml config file and loads the appropriate objects
into the specified environment.
}
\seealso{
Other Logging: 
\code{\link{LogDispatch}},
\code{\link{attach_log_levels}()},
\code{\link{configure_logger}()},
\code{\link{display_log_levels}()},
\code{\link{get_configurations}()},
\code{\link{init_logger}()},
\code{\link{load_log_layouts}()}
}
\concept{Logging}
