% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{sys_context}
\alias{sys_context}
\title{System Context}
\usage{
sys_context()
}
\value{

}
\description{
Wrapper around Sys.info() and \code{get_r_version} that provides
a consolidated list of variables used for logging contexts.
}
