% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{sys_context}
\alias{sys_context}
\title{System Context}
\usage{
sys_context()
}
\value{
system context for evaluating \code{fmt_metric} objects.
}
\description{
Wrapper around Sys.info() and \code{get_r_version} that provides
a consolidated list of variables used for logging contexts.
}
\section{Metrics}{


System Context

\itemize{
\item{"sysname"} : {The operating system name.}
\item{"release"} : {The OS release.}
\item{"version"} : {The OS version.}
\item{"nodename"} : {A name by which the machine is known on the network (if any).}
\item{"machine"} : {A concise description of the hardware, often the CPU type.}
\item{"login"} : {The user's login name, or "unknown" if it cannot be ascertained.}
\item{"user"} : {The name of the real user ID, or "unknown" if it cannot be ascertained.}
\item{"r-ver"} : {R Version in (major).(minor) format.}
}
}

\seealso{
Other Context: 
\code{\link{class_scope}()},
\code{\link{exec_context}()},
\code{\link{format_fn_call}()},
\code{\link{get_call_stack}()},
\code{\link{get_r_version}()},
\code{\link{get_system_info}()}
}
\concept{Context}
