% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{value.new_fmt_log_msg}
\alias{value.new_fmt_log_msg}
\title{Gets the value of a log msg object.}
\usage{
\method{value}{new_fmt_log_msg}(fmt, ...)

\method{value}{new_fmt_log_msg}(fmt, ...)
}
\arguments{
\item{fmt}{object to extract value from.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
object's value

object's value
}
\description{
Gets the value of a log msg object.

Gets the value of a log msg object.
}
