% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{wipe_logger}
\alias{wipe_logger}
\title{Wipe the Logger Instance}
\usage{
wipe_logger()
}
\value{
None.
}
\description{
Cleans up any dangling global instance
from a previous load.
}
\seealso{
Other Configuration: 
\code{\link{apply_active_settings}()},
\code{\link{config_specification}()},
\code{\link{create_log_levels}()},
\code{\link{display_log_levels}()},
\code{\link{ensure_logger}()},
\code{\link{get_active_settings}()},
\code{\link{load_log_layouts}()}
}
\concept{Configuration}
